WIP: refactor: move Exporter import to services and delete main.py
